Remove override from ignores list
