+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/player/chat/ignore/AddIgnore.kt
@@ -6,15 +6,14 @@ import world.gregs.voidps.engine.event.EventDispatcher
 import world.gregs.voidps.engine.event.Events
 
 data class AddIgnore(val name: String) : CancellableEvent() {
-    override val size = 2
+    override val size = 1
 
     override fun parameter(dispatcher: EventDispatcher, index: Int) = when(index) {
         0 -> "add_ignore"
-        1 -> true // prioritise non-overrides
         else -> null
     }
 }
 
-fun ignoresAdd(override: Boolean = true, block: AddIgnore.(Player) -> Unit) {
-    Events.handle("add_ignore", if (override) "*" else true, override = override, handler = block)
+fun ignoresAdd(block: AddIgnore.(Player) -> Unit) {
+    Events.handle("add_ignore", override = false, handler = block)
 }
