+++ game/src/main/kotlin/content/social/ignore/IgnoreList.kts
@@ -1,17 +1,24 @@
 package content.social.ignore
 
+import content.social.chat.privateStatus
+import content.social.friend.friend
+import content.social.friend.world
+import content.social.friend.worldName
 import world.gregs.voidps.engine.client.message
+import world.gregs.voidps.engine.client.updateFriend
+import world.gregs.voidps.engine.data.Settings
 import world.gregs.voidps.engine.data.config.AccountDefinition
 import world.gregs.voidps.engine.data.definition.AccountDefinitions
-import world.gregs.voidps.engine.entity.character.player.Player
+import world.gregs.voidps.engine.entity.character.player.*
 import world.gregs.voidps.engine.entity.character.player.chat.ignore.ignoresAdd
 import world.gregs.voidps.engine.entity.character.player.chat.ignore.ignoresDelete
-import world.gregs.voidps.engine.entity.character.player.name
 import world.gregs.voidps.engine.entity.playerSpawn
 import world.gregs.voidps.engine.inject
+import world.gregs.voidps.network.login.protocol.encode.Friend
 import world.gregs.voidps.network.login.protocol.encode.sendIgnoreList
 
 val accounts: AccountDefinitions by inject()
+val players: Players by inject()
 
 val maxIgnores = 100
 
@@ -19,7 +26,7 @@ playerSpawn { player ->
     player.sendIgnores()
 }
 
-ignoresAdd(override = false) { player ->
+ignoresAdd { player ->
     val account = accounts.get(name)
     if (account == null) {
         player.message("Unable to find player with name '$name'.")
@@ -53,9 +60,13 @@ ignoresAdd(override = false) { player ->
 
     player.ignores.add(account.accountName)
     player.sendIgnore(account)
+    val other = players.get(name)
+    if (other != null && other.friend(player) && !other.isAdmin()) {
+        other.updateFriend(Friend(player.name, player.previousName, world = 0))
+    }
 }
 
-ignoresDelete(override = false) { player ->
+ignoresDelete { player ->
     val accountName = player.ignores.firstOrNull {
         val account = accounts.getByAccount(it) ?: return@firstOrNull false
         name.equals(account.displayName, true) // This packet ignores case for some reason.
@@ -75,6 +86,13 @@ ignoresDelete(override = false) { player ->
 
     name = account.displayName
     player.ignores.remove(account.accountName)
+    if(player.privateStatus != "on") {
+        return@ignoresDelete
+    }
+    val other = players.get(name)
+    if (other != null && (other.friend(player) || other.isAdmin())) {
+        other.updateFriend(Friend(player.name, player.previousName, world = Settings.world, worldName = Settings.worldName))
+    }
 }
